Un-refactor, because it makes better sense that way.
